fix php error and added wp folder name in pre check

+++ includes/functions.php
@@ -188,15 +188,18 @@ if ( ! function_exists( 'instawp_reset_running_migration' ) ) {
 		}
 
 		$instawp_backup_dir = WP_CONTENT_DIR . DIRECTORY_SEPARATOR . INSTAWP_DEFAULT_BACKUP_DIR . DIRECTORY_SEPARATOR;
-		$files_to_delete    = scandir( $instawp_backup_dir );
-		$files_to_delete    = ! is_array( $files_to_delete ) ? array() : $files_to_delete;
-		$files_to_delete    = array_diff( $files_to_delete, array( '.', '..' ) );
+		if ( file_exists( $instawp_backup_dir ) ) {
+			$files_to_delete = scandir( $instawp_backup_dir );
+			$files_to_delete = ! is_array( $files_to_delete ) ? array() : $files_to_delete;
+			$files_to_delete = array_diff( $files_to_delete, array( '.', '..' ) );
 
-		foreach ( $files_to_delete as $file ) {
-			if ( is_file( $instawp_backup_dir . $file ) ) {
-				wp_delete_file( $instawp_backup_dir . $file );
+			foreach ( $files_to_delete as $file ) {
+				if ( is_file( $instawp_backup_dir . $file ) ) {
+					wp_delete_file( $instawp_backup_dir . $file );
+				}
 			}
 		}
+
 		wp_delete_file( ABSPATH . 'fwd.php' );
 		wp_delete_file( ABSPATH . 'dest.php' );
 		wp_delete_file( ABSPATH . 'iwp_log.txt' );
